Changed: Extract downcastTable() and upcastTable() converters to separate files.
